1.5.2 Przeniesienie klasy Node do klasy IntLinkedList.
