+++ src/main/java/edu/kis/vh/nursery/list/IntLinkedList.java
@@ -45,4 +45,37 @@ public class IntLinkedList {
 	int getI() {
 		return i;
 	}
+
+	private static class Node {
+
+		private int value;
+		private Node prev;
+		private Node next;
+
+		protected Node(int i) {
+			value = i;
+		}
+
+		protected int getValue() {
+			return value;
+		}
+
+		protected Node getPrev() {
+			return prev;
+		}
+
+		protected void setPrev(Node prev) {
+			this.prev = prev;
+		}
+
+		protected Node getNext() {
+			return next;
+		}
+
+		protected void setNext(Node next) {
+			this.next = next;
+		}
+	}
+// Komentarz do zadania 'błędy formatowania i konwencji':
+// Kombinacje klawiszy alt + kursor prawo / lewo w Eclipse pozwala na przełączanie między kolejnymi otwartymi plikami 
 }
